Bug fixes and parameter changes

+++ src/common/falsecolor.h
@@ -1,4 +1,4 @@
-/* RCSid $Id: falsecolor.h,v 3.2 2005/11/15 06:52:38 greg Exp $ */
+/* RCSid $Id: falsecolor.h,v 3.3 2005/11/15 20:04:50 greg Exp $ */
 /*
  * Header file for false color tone-mapping.
  *
@@ -53,7 +53,7 @@ fcFixedLog(FCstruct *fcs, double Lwmin, double Lwmax);
  */
 
 extern int
-fcLinearMapping(FCstruct *fcs, TMstruct *tms, int pctile);
+fcLinearMapping(FCstruct *fcs, TMstruct *tms, double pctile);
 /*
 	Compute linear false color map.
 	
@@ -65,7 +65,7 @@ fcLinearMapping(FCstruct *fcs, TMstruct *tms, int pctile);
  */
 
 extern int
-fcLogMapping(FCstruct *fcs, TMstruct *tms, int pctile);
+fcLogMapping(FCstruct *fcs, TMstruct *tms, double pctile);
 /*
 	Compute logarithmic false color map.
